+++ app/package.json
@@ -1,7 +1,7 @@
 {
   "name": "root",
   "version": "1.0.0",
-  "buildVersion": "b003c2e",
+  "buildVersion": "1c77263",
   "private": true,
   "devDependencies": {
     "lerna": "^3.20.2",
