+++ lib/core/src/display/symbol/gpu/geometry.rs
