+++ lib/core/src/display/symbol/gpu/geometry/compound.rs
