+++ lib/core/src/display/symbol/gpu/geometry/primitive.rs
