+++ lib/core/src/display/symbol/gpu/material.rs
