+++ lib/core/src/display/symbol/gpu/registry.rs
