+++ lib/core/src/display/symbol/gpu/shader.rs
