+++ lib/core/src/display/symbol/gpu/shader/builder.rs
