+++ lib/core/src/math.rs
@@ -1,3 +1,6 @@
 //! Root module for math-related utilities.
 
 pub mod topology;
+pub mod types;
+
+pub use types::*;
