+++ lib/core/src/math/types.rs
@@ -1,0 +1,16 @@
+//! This module gathers common math types which are widely used in this project.
+
+pub use nalgebra::Vector2;
+pub use nalgebra::Vector3;
+pub use nalgebra::Vector4;
+
+pub use nalgebra::Matrix2;
+pub use nalgebra::Matrix3;
+pub use nalgebra::Matrix4;
+
+pub use nalgebra::Matrix2x3;
+pub use nalgebra::Matrix2x4;
+pub use nalgebra::Matrix3x2;
+pub use nalgebra::Matrix3x4;
+pub use nalgebra::Matrix4x2;
+pub use nalgebra::Matrix4x3;
