+++ lib/core/src/system/web.rs
@@ -7,3 +7,4 @@ pub use basegl_system_web::*;
 pub use web_sys::HtmlElement;
 pub use web_sys::HtmlDivElement;
 pub use wasm_bindgen::JsValue;
+pub use js_sys::Object;
