+++ lib/core/src/system/web/dom/html.rs
@@ -1,7 +1,5 @@
 //! This module contains all the submodules of the CSS3D rendering system.
 
-mod css3d_object;
-mod css3d_renderer;
+pub mod css3d_renderer;
 
-pub use css3d_object::DomSymbol;
 pub use css3d_renderer::Css3dRenderer;
